Merge my WP Docker script with 10up's image
This allows multiple dev environments to co-exist on the same host.

For a while I've been using a script similar to this to setup a simple dev environment.
Only tested on Linux but should work on OSX

Wanted to have something similar to how VVV works but without using a VM